Explicitly include offical KiCAD libraries used as per-project
KiCAD, like most ECAD packages, includes a set of official part
libraries. KiCAD's default preferences include all of these libraries as
global libraries which makes them available in any project that is
opened. As users may have custom configurations with different sets of
global libraries, explicitly list all of the libraries used by this
project as per-project libraries. For the official libraries, these will
always be available on disk even if they are not in the global library
list.