lpc55_carrier: initial design for LPC55S28JBD64 carrier board
Plugs into LPC55 pin breakout headers on Gemini Bringup to allow
easily changing LPC55 devices. The on-board LPC55 on Gemini Bringup must
not be fitted when using this carrier. Few Gemini Bringups are intended
to be used this way but the ability to do so is vital for testing secure
boot and key provisioning which can easily render a device unusable.