PCB: densification sprint
Most components are now placed in plausible final locations.

About 40% of the board is routed.

Board size is down to about 210x170mm.

Found and fixed a class of errors in the schematic: it turns out that
within one sheet, local labels and hierarchical labels with the same
name are implicitly connected. This caused some of our signals that exit
blocks through e.g. a slew rate control resistor to wind up with both
ends of the resistor connected to the same net!