sp-ethernet: adjust RJ45 LED resistors
LEDs inside Wurth 74980111211 RJ45 have Vf between 1.8V and 2.8V.
Assuming Vf=2.3V is normal, 180R would only provide ~5mA for LEDs
spec'd for 20mA.  56R would be ideal but it'd be the only use of that
value. We use 68R for other LEDs and in this case would provide ~15mA
which is a little low but should be fine for status LEDs.